@@ -14,11 +14,12 @@
 	"dependencies": {
 		"@emotion/react": "^11.9.3",
 		"@emotion/styled": "^11.9.3",
-		"@mui/icons-material": "^5.8.4",
-		"@mui/material": "^5.8.6",
-		"react": "^17.0.2",
-		"react-dom": "^17.0.2",
-		"react-dropzone": "^14.2.2"
+		"@mui/material": "^5.8.7",
+		"chart.js": "^3.8.0",
+		"chartjs-plugin-dragdata": "^2.2.4",
+		"react": "^18.2.0",
+		"react-chartjs-2": "^4.2.0",
+		"react-dom": "^18.2.0"
 	},
 	"devDependencies": {
 		"@babel/core": "^7.18.6",
@@ -26,7 +27,6 @@
 		"@babel/preset-react": "^7.18.6",
 		"babel-loader": "^8.2.5",
 		"css-loader": "^6.7.1",
-		"file-loader": "^6.2.0",
 		"glc-wasm": "file:../glc-rs/pkg",
 		"html-webpack-plugin": "^5.5.0",
 		"style-loader": "^3.3.1",
